Move broadcast_transaction_synchronous to condenser_api #2278
